fix(remotes): split dual-remote wait/merge ops for maintainability
Extract GitHub wait-for-CI + merge I/O into dual-remote-mirror-ops.js so
agent preflight no longer fails source-file baseline growth and launchd
can run --merge-when-ready without human --skip-preflight.
